# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Nhựa API",
        "description": "API quản lý sản phẩm và biến thể sản phẩm",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/product-variants": {
            "get": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Danh sách biến thể sản phẩm",
                "operationId": "3863e44e732ae5f66e7862065318ef37",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Tạo biến thể sản phẩm kèm ảnh",
                "operationId": "faa10e0879e908667a12684e81e5cf5a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductVariantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đã tạo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductVariantResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Tổ hợp không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/product-variants/{id}": {
            "get": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Chi tiết biến thể sản phẩm",
                "operationId": "d6cae9667e44c10c3afefa1cbbc740c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductVariantResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Cập nhật biến thể kèm ảnh",
                "description": "Gửi multipart/form-data bằng POST và đặt _method=PUT.",
                "operationId": "27d83bfc69ded96cf4faf5d524bf7593",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductVariantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đã cập nhật",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductVariantResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    },
                    "422": {
                        "description": "Tổ hợp không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Xóa biến thể sản phẩm",
                "operationId": "c0eb6752fecd75422c926763266afa4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã xóa"
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreProductVariantRequest": {
                "required": [
                    "product_id",
                    "sku",
                    "option_ids"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sku": {
                        "type": "string",
                        "example": "AT-RED-M",
                        "maxLength": 100
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal",
                        "nullable": true,
                        "minimum": 0
                    },
                    "stock": {
                        "type": "integer",
                        "nullable": true,
                        "minimum": 0
                    },
                    "is_active": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "option_ids[]": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "images[]": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "maxItems": 10
                    }
                },
                "type": "object"
            },
            "UpdateProductVariantRequest": {
                "properties": {
                    "_method": {
                        "type": "string",
                        "example": "PUT"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "sku": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "price": {
                        "type": "number",
                        "format": "decimal",
                        "nullable": true,
                        "minimum": 0
                    },
                    "stock": {
                        "type": "integer",
                        "nullable": true,
                        "minimum": 0
                    },
                    "is_active": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "option_ids[]": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "images[]": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "maxItems": 10
                    }
                },
                "type": "object"
            },
...
            "FileResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "file_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "disk": {
                        "type": "string",
                        "example": "public"
                    },
                    "path": {
                        "type": "string",
                        "example": "products/1/image.jpg"
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "image/jpeg"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "type": {
                        "type": "string",
                        "example": "image",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProductVariantResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string",
                        "example": "150000.00",
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FileResource"
                        }
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VariantOptionResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "VariantOptionResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "variant_group_id": {
                        "type": "integer"
                    },
                    "option_code": {
                        "type": "string"
                    },
                    "option_name": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "integer"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer token",
                "scheme": "bearer"
            }
        }
    },
...
}